FaceForPython/requirement.txt

Flask==1.1.1

Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
grpcio-tools==1.25.0


Jinja2==2.10.3
kazoo==2.6.1
MarkupSafe==1.1.1
module==0.0.4
more-itertools==7.2.0
multidict==4.5.2




protobuf==3.10.0
py==1.8.0

pymongo==3.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.5
pytest==5.2.2
pytest-mock==1.11.2
pytest-runner==5.2
python-consul==1.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
semantic-version==2.8.2
six==1.13.0
SQLAlchemy==1.3.11
tenacity==6.0.0
text-unidecode==1.3
tornado==5.1.1
Tornado-MySQL==0.5.1
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
yarl==1.3.0
zipp==0.6.0
